# services/api/src/main.ts
import 'reflect-metadata';
import { NestFactory } from '@nestjs/core';
import { Controller,Module,Get,Post,Param,Query,Req,Body,HttpCode,BadRequestException,NotFoundException,ForbiddenException,UnauthorizedException,ServiceUnavailableException,HttpException } from '@nestjs/common';
import { createHmac,timingSafeEqual,randomUUID } from 'node:crypto';
import { z } from 'zod';
import type { Prisma } from '@prisma/client';
import { db,queue,withDb } from './context.js';
import { identity,subject } from './auth.js';
import { inbound,deploymentInput,transitions } from '../../../packages/contracts/policy.js';
import { assertRuntimeRole } from '../../../packages/db/context.js';
import { ROLES,ROLE_BINDINGS } from '../../../packages/roles/index.js';
import { Commerce } from './commerce.js';
// Money is BIGINT minor units (always < 2^53): serialise as numbers if any slips into a response.
(BigInt.prototype as any).toJSON=function(){return Number(this);};
function parse<T>(schema:z.ZodType<T>,data:unknown):T{const result=schema.safeParse(data);if(!result.success)throw new BadRequestException('Invalid request');return result.data;}
// Every tenant query runs in a transaction bound to the caller's tenant (RLS tenant_isolation).
const as=(who:{tenantId:string;sub:string})=>({tenantId:who.tenantId,userId:who.sub});
@Controller()
class Api {
 @Get('health') health(){return {status:'ok'};}
 @Get('ready') async ready(){try{await db.$queryRaw`SELECT 1`;await (await queue.client).ping();return {status:'ready'};}catch{throw new ServiceUnavailableException();}}
 @Get('v1/roles') roles(){return db.role.findMany({orderBy:{id:'asc'}});}
 // Tenants the caller belongs to (tenant switcher). Memberships are read under the caller's own user scope;
 // each tenant name is then read inside that tenant's context, so nothing outside the caller's tenants is visible.
 @Get('v1/me/memberships') async memberships(@Req() req:any){const sub=await subject(req);const rows=await withDb({userId:sub},tx=>tx.membership.findMany({where:{userId:sub,status:'active'},orderBy:{tenantId:'asc'}}));// tenantStatus lets the switcher show suspended tenants honestly (their runs are suspended by the worker).
 return Promise.all(rows.map(async m=>{const t=await withDb({tenantId:m.tenantId,userId:sub},tx=>tx.tenant.findUnique({where:{id:m.tenantId},select:{name:true,status:true}}));return {tenantId:m.tenantId,role:m.role,name:t?.name??m.tenantId,tenantStatus:t?.status??'unknown'};}));}
 @Post('v1/assessments') async assessment(@Body() body:unknown){const data=parse(z.object({business:z.string().min(1).max(120),need:z.string().min(1).max(1000),consent:z.literal(true)}).strict(),body);return withDb({scope:'public_intake'},tx=>tx.assessment.create({data}));}
 @Post('v1/tenants') async createTenant(@Req() req:any,@Body() body:unknown){const who=await identity(req,undefined,true);if(who.role!=='atlas_owner')throw new ForbiddenException();const data=parse(z.object({slug:z.string().regex(/^[a-z0-9-]+$/).max(80),name:z.string().min(1).max(120)}).strict(),body);return withDb({scope:'platform',userId:who.sub},async tx=>{const tenant=await tx.tenant.create({data:{id:randomUUID(),...data}});await tx.membership.create({data:{tenantId:tenant.id,userId:who.sub,role:'atlas_owner'}});await tx.auditEvent.create({data:{tenantId:tenant.id,actorId:who.sub,action:'tenant.create',objectId:tenant.id}});return tenant;});}
 @Post('v1/tenants/:id/memberships') async membership(@Req() req:any,@Param('id') id:string,@Body() body:unknown){const who=await identity(req,id,true);if(!['atlas_owner','atlas_operator'].includes(who.role))throw new ForbiddenException();const data=parse(z.object({userId:z.string().min(1).max(120),role:z.enum(['tenant_user','tenant_admin'])}).strict(),body);return withDb({tenantId:id,userId:who.sub},async tx=>{const row=await tx.membership.upsert({where:{tenantId_userId:{tenantId:id,userId:data.userId}},create:{tenantId:id,...data},update:{role:data.role}});await tx.auditEvent.create({data:{tenantId:id,actorId:who.sub,action:'membership.set',objectId:row.id}});return row;});}
 @Post('v1/ops/runs/:id/retry') async retry(@Req() req:any,@Param('id') id:string,@Body() body:unknown){const {tenant}=parse(z.object({tenant:z.string()}).strict(),body);const who=await identity(req,tenant,true);if(!who.role.startsWith('atlas_'))throw new ForbiddenException();const run=await withDb(as(who),tx=>tx.taskRun.findFirst({where:{id,tenantId:who.tenantId,state:'dead_letter'}}));if(!run)throw new NotFoundException();if(run.attempts>=9)throw new ForbiddenException('Retry budget exhausted');const job=await queue.getJob(id);if(!job||await job.getState()!=='failed')throw new BadRequestException('Job unavailable');await withDb(as(who),async tx=>{await tx.taskRun.update({where:{id},data:{state:'queued'}});await tx.auditEvent.create({data:{tenantId:who.tenantId,actorId:who.sub,action:'run.retry',objectId:id}});});await job.retry();return {runId:id,state:'queued'};}
 @Get('v1/tenants/:id/deployments') async deployments(@Req() req:any,@Param('id') id:string){const who=await identity(req,id);return withDb(as(who),tx=>tx.deployment.findMany({where:{tenantId:id}}));}
 @Post('v1/deployments') async create(@Req() req:any,@Body() body:unknown){const data=parse(deploymentInput,body);const who=await identity(req,data.tenantId,true);if(!who.role.startsWith('atlas_'))throw new ForbiddenException();const pack=await db.packRelease.findUnique({where:{id:data.packReleaseId}});const role=await db.role.findUnique({where:{id:data.roleId}});if(!await withDb(as(who),tx=>tx.membership.findFirst({where:{tenantId:data.tenantId,userId:data.config.handoff_queue,status:'active',role:{in:['tenant_admin','atlas_operator','atlas_owner']}}})))throw new BadRequestException('Invalid human assignee');const binding=ROLE_BINDINGS[data.roleId];if(!pack||!role||!binding||pack.packId!==binding.packId)throw new BadRequestException('Unsupported role/pack binding');return withDb(as(who),async tx=>{const deployment=await tx.deployment.create({data});for(const toolId of binding.tools)await tx.toolGrant.create({data:{tenantId:data.tenantId,deploymentId:deployment.id,toolId,enabled:true,policy:{requiresApproval:toolId==='agenda.update_status'||ROLES[data.roleId]?.tools[toolId]?.policy==='approval'}}});await tx.auditEvent.create({data:{tenantId:data.tenantId,actorId:who.sub,action:'deployment.create',objectId:deployment.id}});return deployment;});}
 async scopedDeployment(req:any,id:string,write=true){await subject(req);const row=await withDb({scope:'resolve'},tx=>tx.deployment.findUnique({where:{id}}));if(!row)throw new NotFoundException();try{const who=await identity(req,row.tenantId,write);return {row,who};}catch(e){if(e instanceof UnauthorizedException)throw e;throw new NotFoundException();}}
 @Post('v1/deployments/:id/transition') async transition(@Req() req:any,@Param('id') id:string,@Body() body:unknown){const {state}=parse(z.object({state:z.string()}).strict(),body);const {row,who}=await this.scopedDeployment(req,id);if(!who.role.startsWith('atlas_'))throw new ForbiddenException();if(!transitions[row.state]?.includes(state))throw new BadRequestException('Invalid transition');if(['pilot','active'].includes(state))throw new ForbiddenException('Requires homologation; not enabled in Round 1');return withDb({tenantId:row.tenantId,userId:who.sub},async tx=>{const changed=await tx.deployment.updateMany({where:{id,state:row.state},data:{state}});if(!changed.count)throw new BadRequestException('Concurrent transition');await tx.auditEvent.create({data:{tenantId:row.tenantId,actorId:who.sub,action:`deployment.${state}`,objectId:id}});return tx.deployment.findUnique({where:{id}});});}
 @Post('v1/deployments/:id/pause') async pause(@Req() req:any,@Param('id') id:string){return this.transition(req,id,{state:'paused'});}
 @Post('v1/inbound/:provider') @HttpCode(202) async ingest(@Req() req:any,@Param('provider') provider:string,@Body() body:unknown){
 if(provider!=='sandbox')throw new NotFoundException();
 const secret=process.env.INBOUND_HMAC_SECRET;if(!secret||secret.length<32)throw new ServiceUnavailableException();
 const raw=req.rawBody as Buffer;const signature=req.headers['x-atlas-signature'];const expected=createHmac('sha256',secret).update(raw).digest('hex');
 if(typeof signature!=='string'||!/^[a-f0-9]{64}$/.test(signature)||!timingSafeEqual(Buffer.from(signature,'hex'),Buffer.from(expected,'hex')))throw new UnauthorizedException();
 const event=parse(inbound,body);if(Math.abs(Date.now()-Date.parse(event.timestamp))>300000)throw new UnauthorizedException('Stale event');
 // The signed channel binding is the only source of tenant identity for inbound events.
 const binding=await withDb({scope:'inbound'},tx=>tx.channelBinding.findUnique({where:{id:event.channel_binding_id}}));if(!binding||binding.provider!==provider||binding.status!=='active')throw new NotFoundException();
 const run=await withDb({tenantId:binding.tenantId},async tx=>{
 const tenant=await tx.tenant.findUnique({where:{id:binding.tenantId}});if(tenant?.status!=='active')throw new ForbiddenException();
 const deployment=await tx.deployment.findFirst({where:{id:binding.deploymentId,tenantId:binding.tenantId}});if(!deployment)throw new NotFoundException();
 return tx.taskRun.upsert({where:{deploymentId_idempotencyKey:{deploymentId:deployment.id,idempotencyKey:event.event_id}},create:{tenantId:binding.tenantId,deploymentId:deployment.id,externalEventId:event.event_id,idempotencyKey:event.event_id,correlationId:event.event_id,input:event.payload as Prisma.InputJsonValue,state:['sandbox','active','pilot'].includes(deployment.state)?'queued':'suspended'},update:{}});});
 if(run.state==='queued')await queue.add('execute',{runId:run.id},{jobId:run.id,attempts:3,backoff:{type:'exponential',delay:200},removeOnComplete:false,removeOnFail:false});
 return {runId:run.id,state:run.state};
 }
 @Get('v1/runs') async runs(@Req() req:any,@Query('tenant') tenant?:string){const who=await identity(req,tenant);const rows=await withDb(as(who),tx=>tx.taskRun.findMany({where:{tenantId:who.tenantId},orderBy:{startedAt:'desc'},take:100}));return rows.map(({input,...safe})=>safe);}
 @Get('v1/runs/:id') async run(@Req() req:any,@Param('id') id:string,@Query('tenant') tenant?:string){const who=await identity(req,tenant);return withDb(as(who),async tx=>{const row=await tx.taskRun.findFirst({where:{id,tenantId:who.tenantId}});if(!row)throw new NotFoundException();const {input,...safe}=row;return {...safe,events:await tx.taskEvent.findMany({where:{runId:id,tenantId:who.tenantId},orderBy:{createdAt:'asc'}})};});}
 @Get('v1/approvals') async approvals(@Req() req:any,@Query('tenant') tenant?:string){const who=await identity(req,tenant);return withDb(as(who),tx=>tx.approval.findMany({where:{tenantId:who.tenantId},take:100}));}
 @Post('v1/approvals/:id/decide') async decide(@Req() req:any,@Param('id') id:string,@Body() body:unknown){const data=parse(z.object({tenant:z.string(),decision:z.enum(['approved','rejected'])}).strict(),body);const who=await identity(req,data.tenant,true);const approval=await withDb(as(who),tx=>tx.approval.findFirst({where:{id,tenantId:who.tenantId}}));if(!approval)throw new NotFoundException();if(approval.state!=='pending'||approval.expiresAt.getTime()<Date.now())throw new BadRequestException('Expired or resolved');// Approved reschedules (ROLE-001) and approved tool actions (ROLE-002..008) resume the run.
 const resumes=approval.requestedAction==='reschedule'||approval.requestedAction.startsWith('tool:');const updated=await withDb(as(who),async tx=>{const changed=await tx.approval.updateMany({where:{id,tenantId:who.tenantId,state:'pending',expiresAt:{gt:new Date()}},data:{state:data.decision,resolvedAt:new Date(),assigneeId:who.sub}});if(!changed.count)throw new BadRequestException('Already resolved');await tx.auditEvent.create({data:{tenantId:who.tenantId,actorId:who.sub,action:`approval.${data.decision}`,objectId:id}});await tx.taskRun.update({where:{id:approval.runId},data:{state:data.decision==='approved'?(resumes?'queued':'completed'):'cancelled',...(approval.requestedAction!=='reschedule'&&(!resumes||data.decision!=='approved')?{input:{},finishedAt:new Date()}: {})}});return tx.approval.findUniqueOrThrow({where:{id}});});if(data.decision==='approved'&&resumes)await queue.add('execute',{runId:approval.runId},{jobId:`approval-${id}`,attempts:3,backoff:{type:'exponential',delay:200}});return updated;}
 @Get('v1/usage') async usage(@Req() req:any,@Query('tenant') tenant?:string){const who=await identity(req,tenant);return withDb(as(who),tx=>tx.usageRecord.findMany({where:{tenantId:who.tenantId},take:100}));}
 @Get('v1/ops/incidents') async incidents(@Req() req:any,@Query('tenant') tenant?:string){const who=await identity(req,tenant);if(!who.role.startsWith('atlas_'))throw new ForbiddenException();return withDb(as(who),tx=>tx.incident.findMany({where:{tenantId:who.tenantId},take:100}));}
 @Get('v1/ops/metrics') async metrics(@Req() req:any,@Query('tenant') tenant?:string){const who=await identity(req,tenant);if(!who.role.startsWith('atlas_'))throw new ForbiddenException();return {runs:await withDb(as(who),tx=>tx.taskRun.groupBy({by:['state'],where:{tenantId:who.tenantId},_count:true}))};}
}
@Module({controllers:[Api,Commerce]})class AppModule{}
await assertRuntimeRole(db);
const app=await NestFactory.create(AppModule,{rawBody:true,logger:false});
app.useGlobalFilters({catch(exception:unknown,host:any){const status=exception instanceof HttpException?exception.getStatus():500;host.switchToHttp().getResponse().status(status).json({statusCode:status,error:status>=500?'dependency_or_server_error':exception instanceof HttpException?exception.message:'request_rejected'});}});
const limits=new Map<string,{count:number;until:number}>();
app.use((req:any,res:any,next:any)=>{if(req.method!=='POST')return next();const key=req.socket.remoteAddress??'unknown';const now=Date.now();for(const [k,v] of limits)if(v.until<now)limits.delete(k);const entry=limits.get(key)??{count:0,until:now+60000};entry.count++;limits.set(key,entry);if(entry.count>120)return res.status(429).json({error:'rate_limited'});next();});
app.enableShutdownHooks();app.use((req:any,res:any,next:any)=>{res.setHeader('X-Content-Type-Options','nosniff');res.setHeader('Cache-Control','no-store');next();});
await app.listen(Number(process.env.PORT??4000),process.env.BIND_HOST??'127.0.0.1');